pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "canadastats"
version = "0.1.0"
description = "Canada high school sports stats aggregator"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "CanadaStats" }]
license = { file = "LICENSE" }
keywords = ["scraping", "sports", "etl", "sqlite", "streamlit", "canada"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
  "Topic :: Software Development :: Libraries :: Python Modules"
]
dependencies = [
  "beautifulsoup4>=4.12.0",
  "lxml>=5.2.0",
  "pandas>=2.2.0",
  "PyYAML>=6.0.0",
  "python-dateutil>=2.9.0",
  "requests>=2.32.0",
  "streamlit>=1.42.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0"
]

[project.scripts]
canadastats = "canadastats.cli:main"

[project.urls]
Homepage = "https://github.com/<your-username>/<your-repo>"
Issues = "https://github.com/<your-username>/<your-repo>/issues"
Repository = "https://github.com/<your-username>/<your-repo>"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["canadastats*"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
